requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
facenet-pytorch==2.5.3
filelock==3.13.1
Flask==3.0.0
fsspec==2023.10.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
sympy==1.12
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
